+++ packages/core/src/components/ui/Logo/Logo.tsx
@@ -1,12 +1,12 @@
-/* eslint-disable @next/next/no-img-element */
+import styles from './logo.module.scss'
 
 function Logo() {
   return (
-    <img
-      src="/logo.svg"
-      alt="The store logo with with a stylized letter B in a dark blue gradient background"
-      width="176px"
-      height="30px"
+    <span
+      className={styles.fsLogo}
+      data-fs-logo
+      role="img"
+      aria-label="BaseStore logo"
     />
   )
 }