+++ packages/core/src/pages/index.tsx
@@ -82,6 +82,7 @@ function Page() {
         title="Receive our news and promotions in advance. Enjoy and get 10% off on your first purchase."
         actionPath="/"
         actionLabel="Call to action"
+        colorVariant="light"
       />
 
       <Suspense fallback={<ProductShelfSkeleton loading />}>